+++ src/NetCore/Web/Cms/Program.cs
@@ -1,4 +1,4 @@
-using Cms;
+﻿using Cms;
 using Cms.AppCode.Extensions.DependencyInjection;
 using E.Standard.Security.App.Json;
 using E.Standard.Security.Cryptography.Abstractions;
@@ -15,6 +15,8 @@ var builder = WebApplication
 
 #if DEBUG // aspire
 builder.AddServiceDefaults();
+#else
+builder.AddDefaultHealthChecks();
 #endif
 
 var startup = new Startup(builder.Configuration, builder.Environment);
@@ -22,9 +24,10 @@ startup.ConfigureServices(builder.Services);
 
 var app = builder.Build();
 
-#if DEBUG // aspire
+//#if DEBUG // aspire
+// /health & /alive endpoints 
 app.MapDefaultEndpoints();
-#endif
+//#endif
 
 startup.Configure(app,
         app.Environment,